+++ uui/src/lib/IconButton/IconButton.types.d.ts
@@ -1,7 +1,8 @@
 export type IconButtonColor = 'primary' | 'secondary' | 'inherit';
 
 export interface IconButtonProps
-  extends svelte.JSX.HTMLAttributes<HTMLButtonElement | HTMLAnchorElement> {
+  extends svelte.JSX.HTMLAttributes<HTMLButtonElement | HTMLAnchorElement>,
+    Box<HTMLButtonElement | HTMLAnchorElement> {
   toggled?: boolean;
   iconClass?: string;
   toggleable?: boolean;
@@ -10,4 +11,5 @@ export interface IconButtonProps
   iconColor?: IconButtonColor;
   toggledIconContent?: string;
   toggledIconColor?: IconButtonColor;
+  ref?: HTMLButtonElement | HTMLAnchorElement;
 }